requirements.txt
fastapi[all]
aiofiles
asyncpg
python-dotenv
bcrypt
snowflake-id
aioboto3
slowapi
